requirements.txt
openai
python-dotenv
pyyaml
tenacity
typing-extensions
flask==3.0.0